requirements.txt
deepchecks
feature-engine
numpy
pandas
scikit-learn
statsmodels
ydata-profiling
pyclustertend
xgboost
featuretools
dotenv
google-genai